requirement.txt
pytorch==1.13.0
pytorch_geometric==2.1.0
rdkit==22.9.2
wandb==0.13.5
pyemd==0.5.1
easydict==1.9
networkx
matplotlib
pandas
plotly

eden https://github.com/fabriziocosta/EDeN
